Fixed posetracks core initialization for blender 4.3
