+++ blender_manifest.toml
@@ -8,7 +8,7 @@ maintainer = "Ian Sloat"
 type = "add-on"
 website = "https://github.com/yaannsloot/posetracks"
 tags = ["Animation", "Tracking"]
-blender_version_min = "4.3.0"
+blender_version_min = "4.2.0"
 
 license = [
   "SPDX:GPL-3.0-or-later",
