+++ core/module/__init__.py
@@ -18,7 +18,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import os
 import sys
 import json
-import importlib
 import urllib.request
 
 _vmin = sys.version_info.minor
